src/consumer-workload/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
prometheus-client==0.19.0
python-dotenv==1.0.1
pyyaml==6.0.1
psutil==5.9.8
numpy==1.26.4
kafka-python 